+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.1] - 2019-09-25
+### Added
+- A new section, "Other Projects", that currently contains [Spotify Console](https://github.com/danielisgr8/Spotify-Console).
+- Project descriptions now support GitHub repo links.
+
+### Changed
+- Reworked the projects list on the homepage
+  - New sections can now be added much easier.
+
 ## [1.2.0] - 2019-09-24
 ### Changed
 - The body of the homepage is now rendered through React.